Bug 1701425 - Make nsPACMan::mLoadFailureCount atomic, r=necko-reviewers,valentin
Differential Revision: https://phabricator.services.mozilla.com/D133492

UltraBlame original commit: f09cad759c981c192562105e8ea768bf44a490e0

+++ netwerk/base/nsPACMan.cpp
@@ -729,7 +729,7 @@ void nsPACMan::OnLoadFailure() {
   mScheduledReload = TimeStamp::Now() + TimeDuration::FromSeconds(interval);
 
   LOG(("OnLoadFailure: retry in %d seconds (%d fails)\n", interval,
-       mLoadFailureCount));
+       (uint32_t)mLoadFailureCount));
